requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
fpdf
python-dotenv
lap
scipy
